+++ src/Support/macros.php
@@ -40,4 +40,12 @@ function initCarbonMacros()
     Carbon::macro('dateForHumans', fn () => $this->isoFormat('L'));
 
     Carbon::macro('dateTimeForHumans', fn () => $this->isoFormat('L LT'));
+
+    Carbon::macro('timeForHumans', function () {
+        return preg_replace(['/:00 /', '/^0/'], '', $this->isoFormat('LT'));
+    });
+
+    Carbon::macro('randomWorkTime', function (int $from = 9, int $to = 17) {
+        return $this->setTime(rand($from, $to - 1), rand(0, 59), rand(0, 59));
+    });
 }