+++ test.sh
@@ -1,0 +1,3 @@
+#!/bin/env bash
+docker build -t laravel-common-test . \
+  && docker run --rm laravel-common-test