+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.13.0] - 2025-04-17
+
+_Stable release based on [0.13.0-rc.1]._
+
+## [0.13.0-rc.1] - 2025-04-17
+
+### Added
+
+- Add `<x-ig::footer-copy>` blade component.
+- Add `randomWorkTime` macro for `Carbon`.
+- Add `timeForHumans` macro for `Carbon`.
+
+### Changed
+
+- Always show `Ignition` error page in debug mode.
+
+### Fixed
+
+- Translate i18n index page.
+- Fix navig translation for 404 pages.
+
 ## [0.12.1] - 2025-04-16
 
 ### Fixed
@@ -354,6 +375,8 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[0.13.0]: https://https://github.com/internetguru/laravel-common/compare/v0.12.1...v0.13.0
+[0.13.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v0.12.1
 [0.12.1]: https://https://github.com/internetguru/laravel-common/compare/v0.12.0...v0.12.1
 [0.12.0]: https://https://github.com/internetguru/laravel-common/compare/v0.11.2...v0.12.0
 [0.12.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v0.11.2